move set_url_parameter_and_refresh to web_tools
